pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "galsenai-llm"
version = "0.1.0"
description = "Modern CLI-first finetuning project for tool-calling LLMs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.9.0",
  "pyyaml>=6.0.2",
  "rich>=13.9.4",
  "typer>=0.12.5",
]

[project.optional-dependencies]
train = [
  "accelerate>=1.2.1",
  "datasets>=3.2.0",
  "peft>=0.13.2",
  "torch>=2.5.1",
  "transformers>=4.56.0",
  "trl>=0.20.0",
]
unsloth = [
  "unsloth",
]
dev = [
  "pytest>=8.3.4",
  "ruff>=0.8.6",
]

[project.scripts]
galsenai = "galsenai_llm.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/galsenai_llm"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/galsenai_llm/cli.py" = ["B008"]  # typer.Option in defaults is standard Typer pattern

[tool.pytest.ini_options]
pythonpath = ["src"]
